@@ -4,7 +4,7 @@ from pycookiecheat.pycookiecheat import chrome_cookies
 
 __author__ = "Nathan Henrie"
 __email__ = "nate@n8henrie.com"
-__version__ = "v0.4.7"
+__version__ = "v0.5.0"
 
 __all__ = [
     "chrome_cookies",
